Better handling of unknown non-form errors
We have (via `handleErrors`) handled errors against specific form
fields, but if there is an unrecognized error this lets many forms to
either report success, or to silently do nothing.

This updates all the forms using this pattern to instead display a
generic "DAG gone it" generic error in place of the component.