Add a bunch of javascript logic to validate admin status on the current workspace.
This commit has a bunch of stuff in it that gives initial functionality to the User Config pages. I also added some javascript stuff in here to restructure the invite type to match the object structure of the user type. I also reformatted how props are passed down to the child element so that everything will work as expected.